Clean up, rename and delte photofinish
