+++ assets/js/pages/SaptuneDetails/SaptuneDetails.jsx
