+++ assets/js/pages/SaptuneDetails/SaptuneDetails.test.jsx
