+++ assets/js/pages/SaptuneDetails/index.js
